pyproject.toml
[project]
name = "nanobananapro-mcp"
version = "0.1.0"
description = "MCP server for Google Gemini image generation (Nano Banana Pro)"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.2.0",
    "google-genai>=0.8.0",
    "google-api-core>=2.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
nanobananapro-mcp = "nanobananapro_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nanobananapro_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
